Update The Sprawl sheet so rolls show the total rolled, not separate dice (#5968)
* Update Sprawl sheet so rolls show the total rolled, not separate dice

The current sheet requires the players and GM add up the results manually. That's pretty silly, since what you really need to know is the total (there's no significance in the game to the individual die results). The separarate results can still be seen as the tooltip on the roll, like: 2d6+2[edge]+0[forward] = (5+6)+2+0

* Whoops, include the minified code too, not only the pretty source