[Agon Official] Add offical Agon sheet
This is the first version of the offical sheet for Agon by John Harper and Sean Nittner.

There is an older sheet for Agon in the repository, but it's for a very dated development version of the game and is not very useful for the released game. It would probably make sense to deprecate it in favour of this sheet.