RMSS: Pre-upgrade changes (#6148)
More changes setting up the upgrade of the main sheet on the dropdown.  Includes some optimization of the sheetworkers and removal of the now-obsolete hidden inputs which supported the old sheetworkers.  Also fixed the rounding for skill rank calculations and changes the banner graphic source from github instead of imgur.