Five Torches Deep translation capable (#6163)
* Abstract Dungeon

Creation of a sheet for Abstract Dungeon RPG

* Changes needed

Tables removed
CSS reduction
add of a translation term

* Upgraded version

Upgrades :
Added a NPC tab Sheet with roll-template to display a chat review of the NPC
Simplified dice gestion
Added config access out of the game
Improved translation possibilities

* Revert "Upgraded version"

This reverts commit ecb44776153cb72e0c220fda4ecbdbb678cd1c49.

* Update

Upgrades :
Added a NPC tab Sheet with roll-template to display a chat review of the NPC
Simplified dice gestion
Added config access out of the game
Improved translation possibilities

* Update Abstract.JPG

* Update

Hello,
Here is a little update to add the game's name as an input so that people who don't want to see Abstract Dungeon on the sheet while playing Abstract Spandex or whatever can do so ! :)

Also add a little CSS to always get the move/delete options from repeating sections on top of everything.

Cheers :)
Pantoufle

* Update Abstract_HTML.html

* Five Torches Deep FR version creation

This is the creation of a Five Torches Deep character sheet .

French version (english version will follow)

* Five Torches Deep

English version for the FTD sheet :)

I try to upload the images used in the sheet in order to fix the url links in the sheet later. I don't know if it's ok with their current size. Juste let me know :)

* Delete FTD_sheet.json

* fixed image filename

* fix name of preview

* Five Torches Deep Translation capable

The sheet can be translated now, or so I think :D

* change

changes

* FTD sheet ok

ok