Re-edited JSON file to add Preview field
I've added the Preview field.  The image file in the preview field is in the images folder that I've included for this character sheet.  I'm worried that I'm not entering this information properly.  Can you please advise me on what I'm doing wrong?  I'm new to this and need help.