RMSS:  More Pre-Upgrade changes (#6191)
* RMSS: Pre-upgrade changes

More changes setting up the upgrade of the main sheet on the dropdown.  Includes some optimization of the sheetworkers and removal of the now-obsolete hidden inputs which supported the old sheetworkers.  Also fixed the rounding for skill rank calculations and changes the banner graphic source from github instead of imgur.

* Update charsheet_RMSS_fullcalcs_RollTemplates.html

* RMSS- Add change warning

with cut/pasteable link to character sheet forum post to the current sheet and add new picture to update the json file with eventually.

* Update charsheet_RMSS.html

Fix wording on the update warning.