Mongoose Traveller 2e
Streamlined the sheet workers. Lots of people have used the traveller sheets as templates for thier own games, so it would be nice if they had a better example of sheet workers to follow.

Also changed the order of stats in the skill selection dropdowns, to match the official order.