Rewrite entire skill matrix HTML
It was kind of whack since english (and probably other languages) don't
fit in the boxes nicely. I'm still abusing the Hadesology out of the
HTML table element but it looks a bit better. It cuts off long skill
names to make the appearance look cleaner, but that shouldn't be much of
an issue since long skills are identifiable enough.