RMSS:  Pushing new sheet live + minor fixes
Fixed a couple translation key/descriptions that were the wrong way around.  Added a bit more information to the dismissable alert at the top of the sheet.  Added a plea for movement formulas in cm/m if one exists.

Renamed the old sheet and jpg to Old-NoCalc-Sheet, Renamed the new sheet and jpg to replace them.

Updated the JSON file to  have more information about the sheet and added a few sheet setup options to it.