Added new sheet called Paranoia_mongoose_oneclick - allows rolling the dice pool with a single click (#6622)
* New version - allows rolling the dice pool with a single click

The current version of this sheet has a menu system for rolling dice that takes 3 clicks to roll each time. This new sheet has the roll button on the sheet.
Added fill from the left on all checkboxes.
Added space for new skills to be added by the player.
+ a variety of other fixes.

* Re-submiting with trailing comma removed

I've also removed the translation files as you advised translations are submitted through another process.