[Monster of the Week Official] Supplemental to pr/6839 (#6872)
* Monster of the Week Official Sheet

Based on the prior work of Katie Bennie (LethalKitten) and Anistuffs, this is a major revision and interface improvement for the Monster of the Week system.

* Create MotWPreview.jpg

* Proper options

Most <options> weren't implemented properly vis a vis data-i18n due to how the inherited legacy one was implemented. Fixed what I found.

* Update MotWHTML.html

* Update MotWHTML.html

'Enabled' showed up next to IAM flex alt when it was not the flex alt that was advanced, fixed

* Changes to crooked-move1 strings

Preparation for checkboxing the options

* Update translation.json

Lots of surplus spaces. I know it's cosmetic, but I wanted to tighten it up.

* Update MotWHTML.html

Crooked playbook feature adds

* Expert tweak

Expert tweak

* Update MotWHTML.html

Monstrous adjustments

* Spellslinger tweaks

* Update MotWHTML.html

Wronged tweak

* Gumshoe tweaks

* Hex modifications

* Pararomantic tweaks

* Searcher adjustments

* Update MotWHTML.html

Exile tweak

* Update MotWHTML.html

Hardcase

* hardcase, sidekick

* Summoned tweaks

* Update MotWHTML.html

Traveler tweaks

* Update MotWHTML.html

Preparation for rearchitecture — moving sheet workers and roll templates to the top

* Update translation.json

Turns out the Chosen's fate list never had translation tags.

* Big rearchitecture

There were a lot of inefficiencies and repeated code in the prior version. This new version does the same work with about 10,000 fewer lines of HTML, and unlocks a few new features.

* Bugs and Action Science

Fixed some bugs, added the Action Scientist playbook

* Improvements to Mystery Sheet

* Update translation.json

* Update MotWHTML.html

Added in some dynamic bullet points to go with the Hook.

* Update MotWHTML.html

Accidentally got the attribute name wrong on this one

* Update MotWHTML.html

Added in a header

* Update MotWHTML.html

Fixed some issues with Keeper sheet

* Update .gitignore

* Update MotWCSS.css

Spooktacular CSS support

* Update translation.json

Translation support for Spooktacular playtest

* Prelim partial Spooktacular

* Spooktacular playtest playbook added

* Update MotWCSS.css

Roll20 has a bug in how it renders CSS in production. See https://app.roll20.net/forum/post/8867346/production-has-a-css-rendering-bug/ for details. This is hopefully my workaround.

* Weird Move Alt Support

It's now possible to mark in the hunter's reference when you have access to an alternative weird move. Doing so will also allow you to advance that move in the menu that displays when you take a "Advance 2 Basic Moves" improvement.

* Update translation.json

Setting up translation keys for The Constructed

* Update MotWCSS.css

* Constructed playbook added

Added the recently revised Constructed playbook

* Bringing up to date with development copy

* Correcting following merge

* Revert "Correcting following merge"

This reverts commit 8c12e73c923418b68f5262e0ccf735959f2ef3f2.

* Updating the translation files

Looks like they were out of date (and none actually translated for that matter, but that's beyond my scope)

Co-authored-by: Cassie Levett <cassielevett@gmail.com>