Palladium fantasy 1 e (#7013)
* Update palladium_fantasy_1e.html

Added the following tabs: Spells, Psionics, Healer, Wards, Circles, and Configuration.

* Update palladium_fantasy_1e.css

CSS additions for all the html additions.

* Update sheet.json

Updated the instructions line to just point to the wiki page.

* Update palladium_fantasy_1e.html

Added changes to the documentation tab, edited the info link roll button to just point to the Roll20 wiki page.

* Update palladium_fantasy_1e.html

Fixed 2 skill roll buttons.  One didn't have the modifiers inside the ending square brackets, and the other had a dash instead of an underscore in an attribute call.

* Update palladium_fantasy_1e.css

Adjusted size of attack and damage text in spells template (reduced 1.1em to 1em).

* Update palladium_fantasy_1e.css

Added a small amount of padding to the top of the attack and damage segment of the spells roll template.

* Update palladium_fantasy_1e.html

Since I changed how tabs were set up, an existing character sheet would initially only show the tab banner until one of the tabs was selected.  Added a sheetworker to update the current display tab to the new default.