[Chronicles of Darkness] Add Russian translation for Vampire: the Requiem
Added translation for Vampire part of the sheet.

This is a non-official translation, based on the official translation of Vampire: the Masquarade 20th Anniversary Edition, known as "Вампиры: Маскарад, Классические правила" in Russia.
There is no official translation of Vampire: the Requiem, and it appears that there won't be one in the near future.