Update MongooseTraveller.html
Minor tweak to integrate Boon/Bane into a pop-up that handles Boon/Bane logic in-line, removing the existing standalone (persistent) d6 for Boon/Bane.

Short version: rather than having to hover over the results and determine if the Boon/Bane die applies, the player is prompted when rolling from the sheet whether the task roll is normal, with Boon, or with Bane.  From there, Roll20's dice mechanics are used to determine how many dice to roll and which ones to keep.