Create Engine of Life
HTML code for the roll20 Legacy sheet. In this version I included all character playbooks and most of the family playbooks from Legacy: The Engine of Life. The only playbook I left out is the family playbook for The Timestream Refugees. 

Anyone who wants the headache, feel free to add the last family playbook to the code.