Update hu2.html
Added a place to track chi on the character tab, moved the skill level entry field to the same line as the skill name, and fixed a missing class for a combat skill text field that caused it to have a different background color than other input fields.  Updated the documentation and release notes tab.