Update deadlandsclassic.html
Added two common rolls (initiative and perception) to the Core tab.
Updated a few of the skills with concentrations to have their concentrations be separate classes rather than all linked with the same name (language) so they can be added/deleted separately.
Fixed roll name for survival (was performin).
Changed placeholder in Fightin to use (') rather than (`) like the rest of the sheet.
Replaced second instance of TN in arcane abilities with Hand.