[Mothership Official] Add modifier prompt and correct dice type
Three changes have been made to rolls in this sheet. First, modifier prompts are now on all rolls, allowing users to input modifiers for the roll's Target. Second, the dice type has been corrected. Initially it was a d100 (range of 1 to 100), however Mothership uses a d% (range of 0 to 99). This correction manifested as replacing instances of 1d100 with 1d100-1 to get the proper range of values. Finally, the critical successes and critical failures have been updated to reflect the new dice range. CS is now 0 instead of 1, and CF is now 99 instead of 100.