preparing stats for unnat
added a bunch of hidden entries that will represent stat Mod and totals.

Added tooltips to the stat numbers, not just the headers/roll buttons.

Removed a complaint comment because that bug is inconsistent and probably only affected me because I was using Pale Moon. Roll20's latest updates seem to have completely killed support for PM, so now I'm on Vivaldi.

****ed if I'm gonna use Firefox or Chrome.