last min checksz
fixed armor display not displaying total soak
made another attempt at the psyker thing, but it failed again. I went to go phone for help, we'll get em next time. In the meantime, the dropdown works as usual, but DoS adds willpower's unnat at the end. That should fit 80% of people's needs, and we'll have a more permanent fix soon.