Remove mysterious +5 bonus
This looks like a copy/paste issue where line 411 was copied here and Combativeness was changed, but the +5 was left in. There is no +5 to Speed base. Rule book page 215 states: 
```
To know the initial Speed rating of a Character,
make the following addition: Way of
Combativeness + Way of Empathy
```