Update MotWHTML.html
A few small quality of life improvements.

The History & Notes field is now triple in height, so folks will both notice and use it, and also be able to see more of what they put in.

A few of the advances now display a text field when you pick those advances so you can write down what specific choice you made after taking that advance.