Accurate +10 to hit as appropriate
Adds JS that checks for the Accurate prerequisites in Ranged Weapons and Vehicle Ranged Weapons, and appropriately adds +10 to hit.

Moved the Accurate damage DoS dialog to only appear when it's needed

Adds default values to attr_ammo (used for the ammo counter API), and vehicle weapons. [Thanks to ChtiKenny, now I know why default values are a thing.]