New sheet to support The Dragon Prince, Tales of Xadia
This is a new sheet to support the Tales of Xadia game.  It is unofficial and is geared to support what to the best of my ability to guess the actual game will be like once it is released. This contains minimal support for the play test material, but can be done.