Fix quick macros. Stars without number revised
Replaced all instances of "/w" with "/me"

The quick macros in the sheet were not returning anything. I'm no expert on html or roll20 chat commands but discovered this would fix all my problems. Not sure the proper etiquette for contributing on github or anything so sorry if im doing it wrong.