[5eShaped] Fixed broken images
Fixed an issue with the image links of the character sheet using the outdated 'cdn.rawgit.com' instead of 'gistcdn.githack.com'.

It seems like the links for 'cdn.rawgit.com' are starting to shut down, which caused the images to not appear on the character sheet

I also formatted the CSS of the character sheet to be more readable.