Update palladium_fantasy_1e.html
Added a "Total" readonly field and a bonus field for each individual skill in the 4 repeating sections on the skills tab.  
Added a sheetworker to calculate the "Total" based on already existing attributes and the new bonus input.
To make room for those inputs, I made the notes textarea larger and added a checkbox to each skill row to show/hide the notes.