Fixes text color on dark theme in powers
When using a power template, the "emote" part of the power was hard-coded to black. This made it nearly impossible to read on the dark theme. Changing this `.sheet-emote` rule to set color to inherit causes it to inherit the color that normal chat text gets, which is correctly adjusted for the dark theme.