Update blades.css
Added a different value for the --dark-secondarytext in the label styling; the background of some labels on this sheet is the same color as the initial value of --dark-secondarytext so it's making the text invisible. It'll be a much more significant undertaking to modify the sheet to fully take advantage of dark mode, but this should at least allow things to function as they should without full toggle support.