PTUv2: Use strict equality checking for strings on attribute import
Currently when using the import button, all attributes set to 0 will be skipped, because the [`==`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Equality) operator is silly and auto-converts strings into numbers.
 
This is technically a backwards-incompatible change, since this could in theory be used to set some attributes to "please don't override this". I don't think that's useful, however, since simply omitting those attributes achieves the same effect, and if this was intentional I doubt it would be checked this way.

Allowing these values to be changed on the other hand is useful for backup imports and other suchlike.