1 New Feature, 4 Bugfixes
Bugfix - Imports choked if the character in Genesis has a lifestyle that is not associated with a specific SIN, the code is more forgiving for this now.
New Feature - Added Edge Roll buttons to most roll buttons on the sheet.  These can be toggled on/off via the check box in the Options Pane if you don't want them for your character/campaign.
Bugfix - There was an extraneous } character on the D6 roll button, it has been nullified.
Bugfix - Minor style changes to Options Pane and main rolls-computed roll template.
Bugfix - fixed css on PC matrix devices so the notes expand the full width when displayed.