[Snice] Adjusted result text color in dark theme
Adjusted the roll result text color in the Roll20 dark theme. Apparently there was an update not so long ago that made the roll result's text white by default when dark theme is enabled, which then makes it almost invisible on the semi-white backgrounds of the Snice template. Updated the style to set a default text color so it's safe for future updates.