Bugfixes, Images from Github and Darkmode
This in a future proofing update for the Pokeymanz sheet. We cleaned some bugs, changed the images to pull directly from the GitHub folder and added some support for Dark Mode, enough to make the white background border disappear when the user change modes, enhancing the overall experience. We also added a Help Button that explain how some things of the sheet works and point to roll20 wiki for more detailed explanations.