Wrap SotS lifestyle radio buttons into a form
Love this sheet, thanks for making it for our new favorite game!

Lifestyle radio buttons work OK until the sheet is closed. The next time the sheet is opened, it will always choose the final radio button, Opulent, as "on".

Wrapping the radio buttons into a <form> will allow a default check on Comfortable, and maintain the selected value (now an integer) for attr_lifestyle.