Big update for Codex of Worlds
A lot of lines getting swapped out because in the back-end development I switched over to using PUG to continue work on the sheet. This has resulted in a lot of the stuff getting rendered with different whitespace behavior, despite the actual functional HTML content being unchanged. The only actually new stuff here is still a big chunk, as it adds support for all of the Team Playbooks from the upcoming Codex of Worlds supplement (crowdfunding campaign launches mid-February). None of the previously existing functionality of the sheet should be affected, however.