change roll buttons
50% width looks better in the new circular attributes field; not
sure yet whether that should stay for everywhere else in the
character sheet, but let's go with it for now.

I noticed that the "Roll" buttons all had d20 images on them, but
Magical Kitties Save the Day! only uses d6s, so let's switch that
as well.  I had trouble getting the solid or outline d6 icons to
look right, so added a rakish tilt and shifted the icon color to
help there.