move trackables to grid & no spurious spans
I toyed with making the trackables section into a table because
it's a TABLE, but apparently screen readers aren't good with that
and grids are the way to go.

Also noticed that there were a lot of h2 elements with spans
inside them when they could simply be h2 elements (at least as
far as I know), so moved them to only h2 elements for brevity.

Not sure whether this should continue using rem as a form of
measurement, but that's what I used for now.