line text field up with other text fields
This was previously out of line with other text fields in the
sheet; this moves it in some, but feels a little hacky as a way
to achieve it.  I hope to revisit at some point and get it in
line in a better way.

The main challenge seems to be that other text areas are part of
repeating fieldsets, so they're wrapped in multiple elements,
while this particular text field is non-repeating.