switch to grid formatting (work-in-progress)
I may end up reverting this since it's a bigger marshmallow than
I expected, but wanted to try a grid format since it's apparently
the preferred display style.

A lot of TODO items to address still, and am still using the
"flex" names since I haven't fully committed yet.

The bulk of the diffs in the html file are simply from changing
indentation, so look at the diff with "ignore spacing" enabled to
get a better idea of what's changed.

In the html file the changes are mainly removing divs since we
can rely on the grid for element placement and horizontal sizing.
Also switched from span to div in some spots since the elements
seem better-described as divs than spans in the context.

The Trackables section will be a little more complicated, and is
the main reason I'm committing this work-in-progress.  I kind of
want to make it a table, but let's try it as a grid first and see
how it goes.