move trackables for better accessibility
Previously if you tabbed through the fields here, you'd go
through each attribute, then the maximums.  I think the more
intuitive behavior would be attribute, then its maximum, then the
next attribute, so changing to achieve that.

This also allows removing a lot of the wrapping divs, so a lot of
indentation changes.  View diff with spacing changes ignored for
a better understanding of what's actually changed.

The spacing is still a little off on the Trackables section, but
it at least has better alignment with the overall grid now.