replace header image & other small changes
I read somewhere (that I can't find right now) that images on
character sheets are less preferable, so let's try using a CSS
take on the (probably copyrighted) logo the sheet originally had.

This allows us to have a smaller logo, which puts more sheet
content above-the-fold, and also allows us to move (and rename)
the extra-dice-rolls button for even more content above-the-fold.

I read somewhere else that pure black isn't recommended for dark
mode, so let's shift to a slightly-less-dark color for the dark
mode background.  This also brings the textarea backgrounds more
into line with the slight difference between their background and
the sheet background in non-darkmode.