add kitty token, adjust spacing
The change to the roll button at the top of the page also affects
the Attributes roll buttons, so that'll need fixing soon, but I
really like having the character token on the actual character
sheet!

I did ponder whether it should be the avatar instead.  It
probably should.  I'll come back to that at some point.