shift trackables to top-level grid (somewhat)
Get rid of the explicit column definitions for trackables and fix
bug with class definition starting with a number.

Removing the column definitions made the number fields move away
from each other, but I think it's clearer that they're associated
when they're closer, so adjusted margins on them to get them into
the right spot.