tweak "logo" slightly
darkgrey works better as a background color on the drop-shadow
since it shows up in both light and dark modes.  I'm not sure
this is the final version, but I think I like it better than the
previous version.

Also move the overlapping text down slightly--I worry that having
it overlap too much affects legibility for "KITTIES" in the logo.