improve button balance
This fixes the issue with button size in the Attributes section
while also improving the "feel" of the spacing around the button
text.

I'm not sure I have the padding balanced right since the tilted
d6 icon doesn't balance well, but it's better than before.

I'd like to remove the !important, but for now it's needed to get
the changes to show up.  I'm sure there's a better fix, but
working is close enough for now.