move +Add and Modify buttons back to margins
When we switched to grid, the buttons went from sitting at left
and right margins to being bunched up in the first column.
repcontrol class doesn't seem to be defined anywhere, so fine to
create our own, and then we basically copy the fullwide class to
stretch across the entire page.

I'm not sure I like the solution currently given how much blank
space it leaves between the buttons and between elements above
and below, but maybe I just need to remove the <hr> elements to
get that feeling better.