remove non-repeating horizontal line
I'm still grappling with how the <hr> elements should work.

I think they make a lot of sense for the repeating sections where
you want to have a clearer distinction between each item, while
being clear that it's still the same section.

Between sections like this <hr>, it seems clear enough from the
whiteonredtext rows that the section is changing, so leaving out
the line here seems like the best choice to help keep
"above-the-fold" display maximized (though this section usually
is below the fold, so not technically).