adjust whiteonredtext spacing
Previously <br> elements were used for improving spacing between
repeating element blocks, but since those blocks are also divided
by whiteonredtext, let's switch to simply putting a larger margin
on top of whiteonredtext.

This changes the header spacing, which allows things to move more
to a default positioning and have it be a little less cramped.
Had to style the Extra Dice Rolls button directly to get it into
a balanced spot, though.