make darkmode background match sheet color
Currently the darkmode sheet color is a lighter shade than the
black we were using for the background, so there was an unsightly
margin around the character sheet.

I kinda like the darker color, though, so I might look into
overriding it rather than reusing it.  But for now, at least
things match!