remove <h4> tags from repeating element titles
The earlier re-arranging of <h4> tags revealed the
person-problems section wasn't lining up properly, and while
investigating that I realized that the h4 properties weren't
really being applied to the input fields, and instead it was more
like the h4 was being used as a div.

Instead of having a separate div (and needing to propagate the
grid display down to it), I figure it makes more sense to apply
the class directly to the input elements.

It _seems_ like only one of the redtext rules was being applied
to the input fields, so I changed the name of it to more closely
match what it actually is (no red text here), stripped all the
input fields of their h4 tags, and renamed them all to
repeating-title.