* Add roll macro generator
Nested rolls need to be escaped and going beyond three levels of nesting by hand is madness. Also, readjusting the same macro to the current level of nesting poses and extra challenge for misplaced special characters. To minimize time lost finding such bugs, roll macros (for use in roll buttons) can now be generated from two objects: the 'macros' object and the 'reuse' object. The reuse object contains macros (and can contain other macros defined in the same reuse object) to allow easy generation of almost arbitrarily complex nested roll macros. The macros object defines the variables available to the roll and can call reuse macros as well. Combined, both allow to build complex macros in a rather elegant and easy to understand way. The output is the string for use in a roll button. The input sanitation present was written with a user of the roll macro generator in mind, not an intelligent attacker trying to bypass certain checks.