Remove translation folder and <table> element
I just made both corrections.
translation.json replaced the folder and above all I deleted the ugly <table> elements (which actually served no purpose except a structural model for the correct roll template!!)

As I have reworked the character sheet to beautify it in the meantime, there are a lot of changes between the two versions but these are only aesthetic changes. The most profound modification is the embellishment of the checkbox via css to reinforce immersion.

Regarding the French translation, I will actually turn to Crowdin when the sheet is validated.