NPC Defense to PC Defense Style
This is a throwback to the original sheet. There isn't any benefit to using a custom Defense style for the NPCs here. Replaced the NPC Defense sheet with it's PC counterpart.

We also had to fix the Damage Threshold becausewe have weird NPC variables for the NPC character sheet. All of these variables should go away.