HP Max Auto-Set
A small quality of life? A stupid little sheetworker that auto-sets Max HP if it's 0 and you set current health. Trying to gain a little java xp and it is a decent QoL when building a bunch of Characters. It won't mess with Max HP unless it was set to zero and you can increase, but can't decrease below current HP. Haven't done vehicles but works on PCs and NPCs
Fix the HP worker