Reorganized Repeating Attacks
Removed the header. We don't rely on a header to know where to put the things are. Attack/Damage are aligned with a small margin now instead of overlapping.
Moved Target Defense into attack with a drop down box. Moved Attack total to the right. Few minor tweaks to roll template descriptors.