Condition tracker w/ speed
this was painful. still thinking about the reflex def penalty. Probably going to put it in misc and store/save. The only caveat to the CT right now is if you were to increase speed while at ct 10 or helpless, it will break it. It also stores the speed when you open the sheet. I guess there are situations where this will be a bug, but with asynchronous sheetworkers and the need for one checkbox at the time, I don't see a better way without major retooling. Works well other than some pretty edge case situations.