Grapple Mod to sheetworker
So if you have strength selected as the modifier for grapple, if you increase strength, the modifier will not change. That's because the grapple formula pulls grpmod and grpmod is str_max. Rewriting the dropdown to another stat, grptypemode and using a sheetworker to assign grpmod a value on change fixes that.
The benefit to this approach is we're not changing the grapple formula so it won't break things on old sheets, else I would rename all the attributes.