[NEW FEATURE] Dirty System
This sheet update adds the D100 classic version of Dirty System RPG.

All of the different sheets are selectable with a drop-down menu.

Some graphic adjustments were made, as was the change in the link of the images so that they point directly to the Roll20 github.

Following the addition of the Classic D100 version, I had to add entries to the translation file. Not really knowing if I could also make changes in the translation files and not wanting to take the risk of derailing the translation incorporation system that you have set up, I am attaching the gits of the fr.json file up to date.