Submiting upgraded character sheet for Fabula Ultima.
This sheet is an upgraded version of the Fabula Ultima sheet, using the test sheet's as base but with several modifications applied and new content added. The layout was much improved in comparison, it is more user-friendly, more sections were added and the content was reorganized using a better logistic for information.