[Paladin] next small ovehauls
The checkboxes & text in the health section are all unaligned, causing unesthetic alignment in the sheet. By changing the div class, the three lines & corresponding checkboxes realign properly

The "member of the round table" attribute & box is useless - it is a holdover from the old King Arthur Pendragon character sheet on which the Paladin sheet is based. There are no round table knights in the Paladin universe :) However "Paladins" are indeed a thing, and this space is to keep track of the year in which the character potentially receive this super rare honor. It makes sense to drop the old attribute and put the paladin one instead. Impact is extremely minor for users - its just there to track a year, and as I said, it is super rare even for really advanced characters.