NEW SHEET! Hollow Knight TTRPG, Beta Test 1
This is a beta test for the Unofficial HK TTRPG I've been working on over this last month. It's a work in progress, but the people who play this system are excited to finally have a Roll20 character sheet, so I wanted them to have access to what I've done so far. This character sheet accounts for many core aspects of the system, including combat, ability rolls, stamina costs, among other things. This is my first character sheet I've coded, so I hope this beta test is accepted for community use, considering it looks a little jank xD