RTI Rev.3 Hotfix 5-24-24 to remove legacy sanitation
Due to a recent Roll20 update, the legacy sanitation feature was stripping the HTML part of the code from element based styling.
After some reading i have managed to update the sheet so it no longer requires legacy sanitation. (CSE compatible)

This is however a temporary measure since i will try to redesign the sheet to use CSS Grid for layout in the future