[TEN CANDLES] Roll upgrade
Hello, you will find an improvement of a roll by passing it by type=action instead of type=roll.

This allows additional information to be given to players during rolls without them having to calculate them as they had to do with the previous version.

Also updated the translation.json to include the new necessary entries.