update gurps
ENHANCEMENT: Active Defenses. Added a new field for weight of the parrying weapon and tool to calculate weapon breakage. B376.
ENHANCEMENT: Melee Attacks. Added a new field for weight of the attacking weapon and tool to show the parrying weapon's breaking point. B376.
BUGFIX: Fixed the styling for the Duplicate button for melee and ranged attacks.