The Unofficial HKRPG - Beta Test #4 - New Pages and Fixes
This is a beta test for the Unofficial Hollow Knight TTRPG. More pages will be developed in the future. But as of now, I wish to give players access to what I have completed so far, as to see how affective the character sheet is at simplifying gameplay. Original Pull Request was closed for this one, as I found a glitch and fixed it!