Initial Release
This sheet is authorized by the publisher, Dyskami Publishing Company, and developed in cooperation with the owner and creator, Mark MacKinnon.  It supports 3 games based on the Tri-Stat System: Tri-Stat Core, Big Eyes Small Mouth 4E, and Absolute Power. This initial release is considered MVP, a pre-release to gather community feedback and play testing.