Serenity
Added roll buttons to Attributes, Derived Traits and Specialties. Automated the -2 Attribute Penalty and allowed for tabulation of Advancement Points.
This is my first submission so no offense is meant if I've tripped and fallen afoul of rules. This is just a requested update the Serenity sheet on behalf of Leland V. Not a v2 of Serenity. I'm blathering - sorry